#23 move pages to app folder
