+++ src/app/pages/loading/index.js
